### 2021-11-03 5bee9a8
insert cumulative accuracy computation
