### 2021-11-03 #3881 @jbytecode
Dear  @csinva

Thank you for submitting your paper to JOSS.

[ ] Please have a look at the [Reviewer List](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit) and list some suitable potential reviewers here. Please start from bottom of the list. Do not use @ character when you list them - just because we don't want to bother them. If you don't have any suggestions I will - of course - find suitable reviewers for you.

[ ] While I am sending invitations, please correct the missing DOIs that Whedon suggests above.

[ ] You can compile and test your paper using the Whedon utility in page https://whedon.theoj.org/

Thank you in advance.
