+++ references.bib
@@ -5,7 +5,8 @@
   volume={6},
   number={61},
   pages={3192},
-  year={2021}
+  year={2021},
+  doi={10.21105/joss.03192}
 }
 
 @article{van2011numpy,
@@ -16,7 +17,8 @@
   number={2},
   pages={22--30},
   year={2011},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/mcse.2011.37}
 }
 
 @article{mckinney2011pandas,
@@ -27,7 +29,8 @@
   number={9},
   pages={1--9},
   year={2011},
-  publisher={Dallas, TX}
+  publisher={Dallas, TX},
+  doi={10.5281/zenodo.3509134}
 }
 
 @inproceedings{moritz2018ray,
@@ -45,7 +48,8 @@
   volume={41},
   number={4},
   pages={39--45},
-  year={2018}
+  year={2018},
+  doi={10.1145/3399579.3399867}
 }
 
 @article{ivie2018reproducibility,
@@ -56,7 +60,8 @@
   number={3},
   pages={1--36},
   year={2018},
-  publisher={ACM New York, NY, USA}
+  publisher={ACM New York, NY, USA},
+  doi={10.1145/3186266}
 }
 
 @article{saltelli2002sensitivity,
@@ -67,7 +72,8 @@
   number={3},
   pages={579--590},
   year={2002},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/0272-4332.00040}
 }
 
 @article{fisher1937design,
@@ -76,7 +82,8 @@
   journal={The design of experiments.},
   number={2nd Ed},
   year={1937},
-  publisher={Oliver \& Boyd, Edinburgh \& London.}
+  publisher={Oliver \& Boyd, Edinburgh \& London.},
+  doi={10.1038/137252a0}
 }
 
 @article{yu2020veridical,
@@ -87,7 +94,8 @@
   number={8},
   pages={3920--3929},
   year={2020},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+  doi={10.1145/3336191.3372191}
 }
 
 @article{pedregosa2011scikit,
@@ -98,7 +106,8 @@
     pages = {2825--2830},
     year = {2011},
     publisher = {JMLR. org},
-    url = {http://jmlr.org/papers/v12/pedregosa11a.html}
+    url = {http://jmlr.org/papers/v12/pedregosa11a.html},
+    doi={10.5555/1953048.2078195}
 }
 
 @article{rudin2019stop,
