### 2021-11-03 1baae62
remove invalid doi
