+++ references.bib
@@ -106,8 +106,7 @@
     pages = {2825--2830},
     year = {2011},
     publisher = {JMLR. org},
-    url = {http://jmlr.org/papers/v12/pedregosa11a.html},
-    doi={10.5555/1953048.2078195}
+    url = {http://jmlr.org/papers/v12/pedregosa11a.html}
 }
 
 @article{rudin2019stop,
