### 2021-11-03 #3881 @csinva
Thanks for serving as editor!

1. Here are some potential reviewers from the list: richrobe, kmichael08, dionhaefner, vimalthilak,
2. I believe I have corrected the DOIs
3. The paper seems to compile properly

Also tagging my co-authors: @jpdunc23, @rushk014, @aagarwal1996
