### 2021-11-04 #3881 @jbytecode
👋👋👋 Dear @kmichael08  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "_VeridicalFlow: a python package for building trustworthy data-science pipelines with PCS_". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/3881).
