### 2021-11-04 #3881 @jbytecode
@csinva since this is a scientific paper, not a documentation on GitHub, could be please remove the logo in Summary section of the paper?
