### 2021-11-06 #3881 @richrobe
> 👋👋👋 Dear @richrobe 👋👋👋
>
> Would you be willing to assist in reviewing this submission for **JOSS** (**Journal of Open Source Software**)?
>
> JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "_VeridicalFlow: a python package for building trustworthy data-science pipelines with PCS_". You can find more information at the top of this Github issue (#3881).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.

Hi @jbytecode!
Sorry for the late reply, but I'd be happy to review as well!
