### 2021-11-06 #3895 @whedon
[REVIEW]: VeridicalFlow: a python package for building trustworthy data-science pipelines with PCS
**Submitting author:** @csinva (<a href="http://orcid.org/0000-0003-0318-2340">Chandan Singh</a>)
**Repository:** <a href="https://github.com/Yu-Group/veridical-flow" target ="_blank">https://github.com/Yu-Group/veridical-flow</a>
**Version:** v0.1.1
