### 2021-11-09 5ee3d3e
add Vset.predict_with_uncertainties() & update evaluate_uncertainty
- add Vset.predict_with_uncertainties(X_test) to compute mean prediction
probabilities and uncertainties
- update evaluate_uncertainty() to compute cumulative accuracy over MSE between preds and y_real
TODO: compute cumulative accuracy over pred_proba?
