+++ vflow/__init__.py
@@ -4,4 +4,4 @@
 from .vfunc import *
 from .vset import *
 from .pipeline import *
-from .convert import init_args, dict_to_df, compute_interval, evaluate_uncertainty
+from .convert import init_args, base_dict, dict_to_df, compute_interval, evaluate_uncertainty
