### 2021-11-12 c90ca64
Add Vset lazy evaluation
Can also now call a Vset with Vfuncs that take zero args
