### 2021-11-12 70ecdd9
Add optional `keys` arg to Vset.__call__
Using `keys`, `n_out` can now be larger than the number of args passed to
__call__. `key` should be a list of strings of length `n_out` so that the i-th
output dictionary will have keys prepended with `keys[i]`.
