### 2021-11-12 8084b53
Rename "build_Vset" to "build_vset" and add arg "lazy"
