### 2021-11-13 8e503ee
Use subkey.value for entries in dict_to_df
