### 2021-11-13 0b4065c
Add filter_vset_by_metric
filter_vset_by_metric takes in a dictionary with numeric values together with an
arbitrary number of Vsets on the graph to the dictionary and returns copies of
the those Vsets with Vfuncs filtered based on the numeric values in the
dictionary. An intended use-case is to perform a "prediction check" on a
modeling Vset where all but the best performing models are filtered out.
