+++ vflow/helpers.py
@@ -1,8 +1,9 @@
 '''User-facing helper functions included at import vflow
 '''
 
-from vflow.vset import Vset
+from vflow.vset import Vset, PREV_KEY
 from vflow.vfunc import Vfunc
+from vflow.convert import dict_to_df
 
 from itertools import product
 from functools import partial
@@ -46,6 +47,7 @@ def build_vset(name: str, obj, param_dict: dict = {}, *args,
     vfuncs = []
 
     # TODO: better way to check this?
+    # check if obj is a class
     instantiate = isinstance(obj, type)
 
     param_names = list(param_dict.keys())
@@ -55,7 +57,7 @@ def build_vset(name: str, obj, param_dict: dict = {}, *args,
         kwargs_dict = {}
         for param_name, param_val in zip(param_names, tup):
             kwargs_dict[param_name] = param_val
-        # add additional fixed kwargs
+        # add additional fixed kwargs to kwargs_dict
         for k, v in kwargs.items():
             kwargs_dict[k] = v
         if instantiate:
@@ -68,3 +70,66 @@ def build_vset(name: str, obj, param_dict: dict = {}, *args,
     return Vset(name, vfuncs, is_async=is_async,
                 output_matching=output_matching, lazy=lazy,
                 cache_dir=cache_dir, tracking_dir=tracking_dir)
+
+def filter_vset_by_metric(metric_dict: dict, vset: Vset, *vsets: Vset, n_keep: int=1,
+                          bigger_is_better: bool=True, filter_on: list=[],
+                          group: bool=False) -> Vset:
+    '''Returns a new Vset by filtering vset.modules based on values in filter_dict.
+
+    Params
+    -------
+    metric_dict: dict
+        output from a Vset, typically with metrics or other numeric values to use when
+        filtering vset.modules
+    vset: Vset
+        a Vsets
+    *vsets: Vset
+        zero or more additional Vsets
+    n_keep: int (optional)
+        number of entries to keep from vset.modules
+    bigger_is_better: bool (optional)
+        if True, then the top n_keep largest values are retained
+    filter_on: list[str] (optional)
+        if there are multiple metrics in metric_dict, you can specify a subset
+        to consider
+    group: bool (optional)
+        if True, average metrics after grouping values in metric_dict by the
+        input Vset names
+
+    Returns
+    -------
+    *new_vset : Vset
+        Copies of the input Vsets but with Vfuncs filtered based on metrics
+
+    '''
+    df = dict_to_df(metric_dict)
+    vsets = [vset, *vsets]
+    vset_names = []
+    for vset in vsets:
+        if not vset.name in df.columns:
+            raise ValueError(f'{vset.name} should be one of the columns of dict_to_df(metric_dict)')
+        vset_names.append(vset.name)
+    if (len(filter_on) > 0):
+        filter_col = list(metric_dict.keys())[0][-1].origin
+        df = df[df[filter_col].isin(filter_on)]
+    if (group):
+        df = df.groupby(by=vset_names, as_index=False).mean()
+    if bigger_is_better:
+        df = df.sort_values(by='out', ascending=False)
+    else:
+        df = df.sort_values(by='out')
+    df = df.iloc[0:n_keep]
+    for i, vset in enumerate(vsets):
+        vfuncs = vset.modules
+        vfunc_filter = df[vset.name].to_numpy()
+        vfunc_keys = [k for k in vfuncs.keys() if k[0].value in vfunc_filter]
+        new_vfuncs = {k: vfuncs[k] for k in vfunc_keys}
+        new_vset = Vset('filtered_'+vset.name, new_vfuncs, is_async=vset._async,
[…]
