### 2021-11-19 9ac61cd
modify evaluate_uncertainty to convert cumulative mse
