### 2021-11-20 575508a
add optional arg verbose=True to build_vset
If verbose=true names each module as a tuple of elements str("param_name=param_val") for each combination of parameters. Additionally names each Vfunc wrapper with its respective module's tuple.
TODO: Standardize naming schema for Vfunc wrappers across the whole package?
