+++ vflow/subkey.py
@@ -78,7 +78,7 @@ class Subkey:
 
 
     def __repr__(self) -> str:
-        return self.value
+        return str(self.value)
 
 
     def __hash__(self):
