### 2021-11-20 074ead0
fix Vset __init__
Fix bug with build_vset where Subkeys that are themselves tuples won't
merge with non-tuple Subkeys in existing combine_keys logic
Now all Subkeys (regardless of whether they are tuples) are wrapped in
trailing tuples
