### 2021-11-20 b7bb9c1
fix test_pipelines tests that use build_vset
TODO: Should build_vset verbose default True or False?
