### 2021-11-21 Yu-Group/veridical-flow#14 @kmichael08
Notebook example – no data given in the link
I run the examples, and the example below uses the data not uploaded to the repo.

https://github.com/Yu-Group/veridical-flow/blob/cfa83c14a7cc2673f9db8acd593072c9c321b609/notebooks/01_enhancer.ipynb?short_path=14469c6#L69

Due to JOSS submission:  https://github.com/openjournals/joss-reviews/issues/3895
