### 2021-11-21 Yu-Group/veridical-flow#15 @kmichael08
Torch required to run an example notebook
In one of the example notebooks, you use Torch, which is not in the requirements.
Perhaps, you could add that to dev requirements or create a separate file with extra requirements to run examples.
Ideally, also describe the process of running the examples and the need to install extra libraries.

https://github.com/Yu-Group/veridical-flow/blob/cfa83c14a7cc2673f9db8acd593072c9c321b609/notebooks/03_computer_vision_dnn.ipynb?short_path=4e5cd33#L26

Due to JOSS submission: https://github.com/openjournals/joss-reviews/issues/3895
