### 2021-11-21 Yu-Group/veridical-flow#16 @kmichael08
Redundant parentheses
https://github.com/Yu-Group/veridical-flow/blob/cfa83c14a7cc2673f9db8acd593072c9c321b609/vflow/convert.py#L247

Also in some other places in the code.

Due to JOSS submission: openjournals/joss-reviews#3895
