### 2021-11-21 Yu-Group/veridical-flow#17 @kmichael08
Too broad exception clause
Please specify the exception, you should control what exceptions can be expected here

https://github.com/Yu-Group/veridical-flow/blob/cfa83c14a7cc2673f9db8acd593072c9c321b609/vflow/pipeline.py#L39

Due to JOSS submission: openjournals/joss-reviews#3895
