### 2021-11-21 Yu-Group/veridical-flow#18 @kmichael08
Redundant requirement
It seems you're not using sklearn anywhere in the package.
If that's the case and it is used for examples, perhaps you could create a separate requirements set and keep the original as slim as possible.

https://github.com/Yu-Group/veridical-flow/blob/cfa83c14a7cc2673f9db8acd593072c9c321b609/setup.py#L26
Due to JOSS submission: openjournals/joss-reviews#3895
