### 2021-11-21 #3895 @kmichael08
Hey @csinva, @jbytecode. I went through the code, paper and documentation.
VeridicalFlow seems to be an exciting and solid package, and I could use it in the future!
I list some issues below, the most pressing for me are the issues with examples.

### Software

- Smooth setup
- Minor comment for requirements (https://github.com/Yu-Group/veridical-flow/issues/18)
- Code looks good in general, although I found some minor issues. I marked them as issues in the repo. I recommend you to run Linter or some other tool (I briefly scanned it with PyCharm's inspection) to find all occurrences of these problems and eliminate warnings, style issues, typos etc.

### Tests

- Tests generally look good, and they are run automatically with GitHub Actions.
- I created minor issues related to tests.

### Documentation

- Looks neat; I added one minor issue.
- You could extend contributing guidelines to aid users seeking support or trying to add an issue. For instance, you could think of Issue templates (e.g. see Issues in pandas).
- When giving an example in the docs and readme, you could attach the output as well to aid understanding.

### Examples
- I managed to run the first example so far. I addressed some problems in the issues. It seems some data to run the examples is missing. Also, there's a minor problem with the required libraries. I'm not sure about the examples in subdirectories (ref, experimental). Are they ready to be used? Perhaps you could describe the examples a little bit?
- Maybe I don't see it, but you could demonstrate computability and reproducibility in the examples (Ray and MLFlow)?

### Paper

- Overall, the paper has a good structure
- However, from the paper itself, I do not fully understand the functionality of the package. I'd suggest adding an example to the paper. Ideally, along with the output like in the first example (only that helped me understand what it does).
- Additionally, could you elaborate on what packages does it compare to and how is it better? My current understanding is that the package functionality is related to sklearn's Pipeline, Tensorflow compute graph and potentially some AutoML solutions.
- Could you elaborate on the stability and how is it supported by the package?
- Perhaps you could introduce some significant concepts from this work, such as VSet?

### Paper – language

- Keep in mind, I'm not a native speaker, so use your judgement and feel free to ignore my comments. The number of lines given for each example:
- 10: provides a -> provides with a
- 17: redundant comma
- 25: redundant comma
- 41-43: a bit unclear sentence
- Finally, I find it strange to write: "Data-practitioners", "machine-learning", "data-science". I always see them without a hyphen

I hope this will be helpful!
I'm ready to accept after running some more examples.
