### 2021-11-21 bb3a05b
remove redundant parentheses
