### 2021-11-30 39e37c5
Fix incompatibility between filter_vset_by_metric and build_vset
