### 2021-12-01 f73857f
Rework PREV_KEY and pipeline graph construction
This fixes an issue where re-using a Vset (e.g., an evaluation Vset) would cause
build_graph to fail due to a self reference in the Vset's PREV_KEY attr. Now,
only Vset.fit modifies the Vset's PREV_KEY, but output dictionaries from other
methods (predict, transform, __call__, etc.) do contain the full PREV_KEY
history.

For dictionaries, PREV_KEY is now a nested tuple like (vset1, arg1_1, arg1_2,
...), where the first entry is a reference to the vset that created the
dictionary and the following entries are similar tuples like (vset2, arg2_1,
...
[…]
