### 2021-12-02 f10cc28
add dict_data, dict_keys helper functions
