### 2021-12-02 fefe48e
github CI use dev reqs
