+++ .github/workflows/python-package.yml
@@ -22,7 +22,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install . --no-cache-dir
+          pip install .[dev] --no-cache-dir
       - name: Test with pytest
         run: |
           pytest
