### 2021-12-02 Yu-Group/veridical-flow#18 @csinva
Thanks for pointing this out! We've moved it into the optional "dev" requirements in the [setup.py](https://github.com/Yu-Group/veridical-flow/blob/master/setup.py) file. (So it would be installed used `pip install .[dev]` rather than the standard `pip install .`).
