### 2021-12-02 fe41882
update notebooks deps
