### 2021-12-02 Yu-Group/veridical-flow#15 @csinva
Thanks for this recommendation! I added an extra requirement section in the `setup.py` file specifically for notebook examples (and also added a [note in the readme](https://github.com/Yu-Group/veridical-flow#documentation) + a [note in the notebook](https://github.com/Yu-Group/veridical-flow/blob/master/notebooks/03_computer_vision_dnn.ipynb)).
