+++ vflow/convert.py
@@ -1,22 +1,24 @@
-'''Useful functions for converting between different types (dicts, lists, tuples, etc.)
-'''
+"""Useful functions for converting between different types (dicts, lists, tuples, etc.)
+"""
 from copy import deepcopy
+from typing import Union
 from uuid import uuid4
 
-from vflow.vset import PREV_KEY
-from vflow.vfunc import VfuncPromise
-from vflow.subkey import Subkey
 import pandas as pd
 from pandas import DataFrame
 
+from vflow.subkey import Subkey
+from vflow.vfunc import VfuncPromise
+from vflow.vset import PREV_KEY
 
-def init_args(args_tuple: tuple, names=None):
-    ''' converts tuple of arguments to a list of dicts
+
+def init_args(args_tuple: Union[tuple, list], names=None):
+    """ converts tuple of arguments to a list of dicts
     Params
     ------
     names: optional, list-like
         gives names for each of the arguments in the tuple
-    '''
+    """
     if names is None:
         names = ['start'] * len(args_tuple)
     else:
@@ -25,66 +27,75 @@ def init_args(args_tuple: tuple, names=None):
     output_dicts = []
     for (i, ele) in enumerate(args_tuple):
         output_dicts.append({
-            (Subkey(names[i], 'init'), ): args_tuple[i],
-            PREV_KEY: ('init', ),
+            (Subkey(names[i], 'init'),): args_tuple[i],
+            PREV_KEY: ('init',),
         })
     return output_dicts
 
 
 def s(x):
-    '''Gets shape of a list/tuple/ndarray
-    '''
+    """Gets shape of a list/tuple/ndarray
+    """
     if type(x) in [list, tuple]:
         return len(x)
     else:
         return x.shape
+
 
 def init_step(idx, cols):
     for i in range(idx, len(cols)):
         if cols[i] != 'init':
             return 'init-' + cols[i]
 
+
 def dict_to_df(d: dict, param_key=None):
-    '''Converts a dictionary with tuple keys
+    """Converts a dictionary with tuple keys
     into a pandas DataFrame, optionally seperating
     parameters in param_key if not None
-    '''
-    d_copy = {tuple([sk.value for sk in k]):d[k] for k in d if k != PREV_KEY}
+    """
+    d_copy = {tuple([sk.value for sk in k]): d[k] for k in d if k != PREV_KEY}
     df = pd.Series(d_copy).reset_index()
     if len(d_copy.keys()) > 0:
         key_list = list(d.keys())
         subkey_list = key_list[0] if key_list[0] != PREV_KEY else key_list[1]
         cols = [sk.origin for sk in subkey_list] + ['out']
         # set each init col to init-{next_module_set}
-        cols = [c if c != 'init' else init_step(idx, cols) for idx, c in enumerate(cols) ]
+        cols = [c if c != 'init' else init_step(idx, cols) for idx, c in enumerate(cols)]
         df.set_axis(cols, axis=1, inplace=True)
         if param_key:
             df = df.join(pd.DataFrame(df[param_key].tolist()))
     return df
 
-def compute_interval(df: DataFrame, d_label, wrt_label, accum: list=['std']):
-    '''Compute an interval (std. dev) of d_label column with 
+
+def compute_interval(df: DataFrame, d_label, wrt_label, accum=None):
+    """Compute an interval (std. dev) of d_label column with
     respect to pertubations in the wrt_label column
-    '''
+    """
+    if accum is None:
+        accum = ['std']
     df = df.astype({wrt_label: str})
     return df[[wrt_label, d_label]].groupby(wrt_label).agg(accum)
 
[…]
