+++ tests/test_pipelines.py
@@ -1,11 +1,10 @@
-from functools import partial
-
 import time
+from functools import partial
+from shutil import rmtree
+
 import numpy as np
 import pandas as pd
-import pytest
 import sklearn
-from shutil import rmtree
 from numpy.testing import assert_equal
 from sklearn.datasets import make_classification
 from sklearn.ensemble import RandomForestRegressor
@@ -17,9 +16,9 @@ from sklearn.tree import DecisionTreeClassifier, DecisionTreeRegressor
 from sklearn.utils import resample
 
 from vflow import Vset, init_args, build_vset  # must install vflow first (pip install vflow)
-from vflow.vset import PREV_KEY
 from vflow.pipeline import build_graph
 from vflow.subkey import Subkey as sm
+from vflow.vset import PREV_KEY
 
 
 class TestPipelines:
@@ -28,8 +27,8 @@ class TestPipelines:
         pass
 
     def test_subsampling_fitting_metrics_pipeline(self):
-        '''Simplest synthetic pipeline
-        '''
+        """Simplest synthetic pipeline
+        """
         # initialize data
         np.random.seed(13)
         X, y = sklearn.datasets.make_classification(n_samples=50, n_features=5)
@@ -47,7 +46,7 @@ class TestPipelines:
         # fit models
         modeling_set = Vset(name='modeling',
                             modules=[LogisticRegression(max_iter=1000, tol=0.1),
-                                          DecisionTreeClassifier()],
+                                     DecisionTreeClassifier()],
                             module_keys=["LR", "DT"])
 
         modeling_set.fit(X_trains, y_trains)
@@ -71,8 +70,8 @@ class TestPipelines:
         assert len(hard_metrics.keys()) == 13
 
     def test_feat_engineering(self):
-        '''Feature engineering pipeline
-        '''
+        """Feature engineering pipeline
+        """
         # get data as df
         np.random.seed(13)
         data = sklearn.datasets.load_boston()
@@ -83,9 +82,11 @@ class TestPipelines:
                                                      names=['X_train', 'X_test', 'y_train', 'y_test'])
 
         # feature extraction - extracts two different sets of features from the same data
-        def extract_feats(df: pd.DataFrame, feat_names=['CRIM', 'ZN', 'INDUS', 'CHAS']):
-            '''extract specific columns from dataframe
-            '''
+        def extract_feats(df: pd.DataFrame, feat_names=None):
+            """extract specific columns from dataframe
+            """
+            if feat_names is None:
+                feat_names = ['CRIM', 'ZN', 'INDUS', 'CHAS']
             return df[feat_names]
 
         feat_extraction_funcs = [partial(extract_feats, feat_names=['CRIM', 'ZN', 'INDUS', 'CHAS']),
@@ -96,7 +97,6 @@ class TestPipelines:
                                output_matching=True)
 
         X_feats_train = feat_extraction(X_train)
-        X_feats_test = feat_extraction(X_test)
 
         modeling_set = Vset(name='modeling',
                             modules=[DecisionTreeRegressor(), RandomForestRegressor()],
@@ -118,19 +118,20 @@ class TestPipelines:
         # inspect the pipeline
         # for k in hard_metrics:
         #     print(k, hard_metrics[k])
-        G = build_graph(hard_metrics, draw=True)
+        _ = build_graph(hard_metrics, draw=True)
 
         # asserts
-        k1 = (sm('X_train', 'init'), sm('feat_extraction_0', 'feat_extraction', True), sm('X_train', 'init'), sm('y_train', 'init'),
[…]

[…]
