### 2021-12-07 6572690
improve coverage && test warnings
add test_convert tests for dict_to_df(), compute_interval()
and build_graph from test_feat_engineering()
