### 2021-12-07 1280836
Updates to paper based on @kmichael08's review
