+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'VeridicalFlow: a python package for building trustworthy data-science pipelines with PCS'
+title: 'VeridicalFlow: a Python package for building trustworthy data-science pipelines with PCS'
 tags:
   - python
   - stability
@@ -33,48 +33,198 @@ bibliography: references.bib
 # Summary
 
 `VeridicalFlow` is a Python package for simplifying building reproducible and trustworthy data-science pipelines using the PCS framework [@yu2020veridical].
-It provides users a simple interface for stability analysis, i.e. checking the robustness of results from a data-science pipeline to various judgement calls made during modeling.
-This ensures that arbitrary judgement calls made by data-practitioners (e.g. specifying a default imputation strategy) do not dramatically alter the final conclusions made in a modeling pipeline.
-In addition to wrappers facilitating stability analysis, `VeridicalFlow` also automates many cumbersome coding aspects of python pipelines, including experiment tracking and saving, parallelization, and caching, all through integrations with existing python packages.
+It provides users with a simple interface for stability analysis, i.e. checking the robustness of results from a data-science pipeline to various judgement calls made during modeling.
+This ensures that arbitrary judgement calls made by data practitioners (e.g. specifying a default imputation strategy) do not dramatically alter the final conclusions made in a modeling pipeline.
+In addition to wrappers facilitating stability analysis, `VeridicalFlow` also automates many cumbersome coding aspects of Python pipelines, including experiment tracking and saving, parallelization, and caching, all through integrations with existing Python packages.
 Overall, the package helps to code using the PCS (predictability-computability-stability) framework, by screening models for predictive performance, helping automate computation, and facilitating stability analysis.
 
 # Statement of need
 
-Predictability, computability, and stability are central concerns in modern statistical/machine-learning practice, as they are required to help vet that findings reflect reality, can be reasonably computed, and are robust as the many judgement calls during the data-science life cycle which often go unchecked [@yu2020veridical]. 
+Predictability, computability, and stability are central concerns in modern statistical/machine learning practice, as they are required to help vet that findings reflect reality, can be reasonably computed, and are robust as the many judgment calls during the data science life cycle which often go unchecked [@yu2020veridical]. 
 
-The package focuses on stability, but also provides wrappers to help support and improve predictability and computability.
+The package focuses on stability but also provides wrappers to help support and improve predictability and computability.
 Stability is a common-sense principle related to notions of scientific reproducibility [@fisher1937design,@ivie2018reproducibility], sample variability, robust statistics, sensitivity analysis [@saltelli2002sensitivity], and stability in numerical analysis and control theory.
 Moreover, stability serves as a prerequisite for understanding which parts of a model will generalize and can be interpreted [@murdoch2019definitions].
 
 Importantly, current software packages offer very little support to facilitate stability analyses.
 `VeridicalFlow` helps fill this gap by making stability analysis simple, reproducible, and computationally efficient.
-This enables a practitioner to represent a pipeline with many different perturbations in a simple-to-code way, while using prediction analysis as a reality check to screen out poor models. 
+This enables a practitioner to represent a pipeline with many different perturbations in a simple-to-code way while using prediction analysis as a reality check to screen out poor models. 
 
 # Features
 
-Using `VeridicalFlows`'s simple wrappers easily enables many best practices for data science, and makes writing pipelines easy.
[…]
