### 2021-12-09 6d9b4db
Add 'reps' argument to build_vset
