### 2021-12-09 d9cb2d5
Add fit_transform method to Vset, and allow for refitting
