### 2021-12-09 7e63466
Add FILTER_PREV_KEY to Vset to make correct PREV_KEY when refitting
