### 2021-12-09 7938c86
Handle param_key='out' case in dict_to_df
