+++ notebooks/readme.md
@@ -1,0 +1,9 @@
+# notebooks
+
+This folder contains notebooks showing demos for various features.
+
+This folder contains demos for common features.
+
+The [computational](computational) folder contains demos showing how to integrate with mlflow/ray for computational speedups.
+
+The [experimental](experimental) folder contains demos for experimental features which are not well-tested yet.
