### 2021-12-10 2e04c6a
Handle DataFrame columns with list/np.ndarray in perturbation_stats
