+++ vflow/convert.py
@@ -4,6 +4,7 @@ from copy import deepcopy
 from typing import Union
 from uuid import uuid4
 
+import numpy as np
 import pandas as pd
 from pandas import DataFrame
 
@@ -91,16 +92,73 @@ def compute_interval(df: DataFrame, d_label, wrt_label, accum=None):
     return df[[wrt_label, d_label]].groupby(wrt_label).agg(accum)
 
 
-def perturbation_stats(df: DataFrame, *groups: str, wrt_col: str = 'out',
-                       func=None):
+def perturbation_stats(df: DataFrame, *groups: str, wrt_col: str='out',
+                       func=None, prefix: str=None, split: bool=False):
     """Compute statistics for wrt_col in df, conditional on groups
+
+    Params
+    ------
+    df: pandas.DataFrame
+        DataFrame on which to compute statistics.
+    *groups: str
+        Columns names in `df` to group on.
+    wrt_col: str (optional)
+        Column name in `df` on which to compute statistics. Defaults to `'out'`.
+    func: function, str, list or dict (optional), default None
+        A list of functions or function names to use for computing
+        statistics, analogous to the parameter of the same name in
+        pandas.core.groupby.DataFrameGroupBy.aggregate. If `None`, defaults to
+        `['count', 'mean', 'std']`.
+    prefix: str (optional), default None
+        A string to prefix to new columns in output DataFrame. If `None`,
+        uses the value of `wrt_col`.
+    split: bool (optional), default False
+        If `True` and `df[wrt_col]` has `list` or `numpy.ndarray` entries, will
+        attempt to split the entries into multiple columns for the output.
     """
     if func is None:
         func = ['count', 'mean', 'std']
+    if prefix is None:
+        prefix = wrt_col
     groups = list(groups)
-    df = df.groupby(groups).agg(func)[wrt_col]
-    df.reset_index(inplace=True)
-    return df.sort_values(groups[0])
+    gb = df.groupby(groups)[wrt_col]
+    mean_or_std = type(func) is list and 'mean' in func or 'std' in func
+    list_or_ndarray = type(df[wrt_col].iloc[0]) in [list, np.ndarray]
+    if mean_or_std and list_or_ndarray:
+        dfs = [gb.get_group(grp) for grp in gb.groups]
+        wrt_arrays = [np.stack(d.tolist()) for d in dfs]
+        n_cols = wrt_arrays[0].shape[1]
+        df_out = pd.DataFrame(gb.agg('count'))
+        df_out.columns = [f'{prefix}-count']
+        if 'mean' in func:
+            if split:
+                col_means = [arr.mean(axis=0) for arr in wrt_arrays]
+                col_names = [f'{prefix}{i}-mean' for i in range(n_cols)]
+                wrt_means = pd.DataFrame(col_means, columns=col_names,
+                                         index=gb.groups.keys())
+            else:
+                col_means = [{f'{prefix}-mean': arr.mean(axis=0)} for arr in wrt_arrays]
+                wrt_means = pd.DataFrame(col_means, index = gb.groups.keys())
+            wrt_means.index.names = df_out.index.names
+            df_out = df_out.join(wrt_means)
+        if 'std' in func:
+            if split:
+                col_stds = [arr.std(axis=0, ddof=1) for arr in wrt_arrays]
+                col_names = [f'{prefix}{i}-std' for i in range(n_cols)]
+                wrt_stds = pd.DataFrame(col_stds, columns=col_names,
+                                        index=gb.groups.keys())
+            else:
+                col_stds = [{f'{prefix}-std': arr.std(axis=0, ddof=1)} for arr in wrt_arrays]
[…]
