### 2021-12-10 acdee22
Updates to paper including working examples
