+++ paper.md
@@ -53,7 +53,7 @@ This enables a practitioner to represent a pipeline with many different perturba
 # Features
 
 Using `VeridicalFlows`'s simple wrappers easily enables many best practices for
-data science, and makes writing pipelines easy.
+data science and makes writing powerful pipelines straightforward.
 
 | Stability                                                    | Computability                                                | Reproducibility                          |
 | ------------------------------------------------------------ | ------------------------------------------------------------ | ---------------------------------------- |
@@ -66,9 +66,9 @@ reasonable or realistic perturbations. The central concept is the `Vset`, short
 for "veridical set", which replaces a given static pipeline step with a set of
 functions subject to different pipeline perturbations that are documented and
 argued for via PCS documentation [@yu2020veridical]. Then, a set of useful
-analysis functions and computations enable easily assessing the stability to
-these perturbations on top of predictive screening for reality checks to filter
-unstable pipeline paths from further analysis.
+analysis functions and computations enable simple assessment of the pipeline's
+stability to these perturbations on top of predictive screening for reality
+checks to filter unstable pipeline paths from further analysis.
 
 ## A stability analysis example
 
@@ -82,18 +82,23 @@ calculates the permutation importance metric via the function
 `sklearn.inspection.permutation_importance`.
 
 ```python
-from vflow import Vset
+from vflow import Vset, build_vset
+from sklearn.impute import KNNImputer, SimpleImputer
 from sklearn.inspection import permutation_importance
 
-# define preprocessing functions preproc_func{1,2,3}
-def preproc_func1(X, y):
-    ...
+preproc_list = [SimpleImputer(strategy='mean'),
+                SimpleImputer(strategy='median'),
+                KNNImputer()]
 
-# create a Vset which varies over the list of preprocessing functions
-preproc_set = Vset("preproc", [preproc_func1, preproc_func2, preproc_func3])
+# create a Vset which varies over preproc_list
+# we use output_matching=True to ensure that preprocessing strategies
+# match throughout the pipeline
+preproc_set = Vset("preproc", preproc_list, ['mean', 'med', 'knn'],
+                   output_matching=True)
 
-# create the feature importance Vset
-feat_imp_set = Vset('feat_imp', permutation_importance)
+# create the feature importance Vset using helper build_vset
+feat_imp_set = build_vset('feat_imp', permutation_importance,
+                          n_repeats=4)
 ```
 
 ### 2. Define model hyperparameter perturbations
@@ -104,23 +109,22 @@ helper `build_vset` to create hyperparameter perturbations for random forest.
 
 ```python
 from sklearn.ensemble import RandomForestRegressor as RF
-from vflow import build_vset
 
 # hyperparameters to try
 RF_params = {
     'n_estimators': [100, 300],
-    'min_samples_split': [2, 10],
-    'max_features': ['sqrt', 'log2']
+    'min_samples_split': [2, 10]
 }
 
-# we could instead pass a list of distinct models and corresponding param dicts
+# we could instead pass a list of distinct models
+# and corresponding param dicts
 RF_set = build_vset('RF', RF, RF_params)
 ```
 
 ### 3. Define data perturbations
 
 For stability analysis, it is often useful to add data perturbations such as the
-bootstrap in order to average over resampling variability in the data.
+bootstrap in order to assess stability over resampling variability in the data.
 
 ```python
 from sklearn.utils import resample
@@ -132,26 +136,26 @@ boot_set = build_vset('boot', resample, reps=100, lazy=True)
 
[…]
