### 2021-12-11 #3895 @csinva
Thanks again for all your thoughtful feedback @kmichael08. We’ve gone through and applied all of your recommendations resulting in a much-improved package / paper!

We’ve responded to [each of the issues](https://github.com/Yu-Group/veridical-flow/issues?q=is%3Aissue+is%3Aclosed) individually in their own threads and improved the package accordingly. We’ll also mention some of the things you brought in the review comment above:

- We've added a fair amount to the paper, especially an example of using the package and its outputs.
- We cleaned up the example notebooks + set up the notebooks for demonstrating computability (mlflow and ray) more prominently in the `notebooks/computational` folder
- We went over the code with PyCharm’s inspection as well and fixed the style-issues present in the code

Sincerely,
Chandan
