### 2021-12-11 67351fd
update contributing.md
following pandas contributing guide
