### 2021-12-12 Yu-Group/veridical-flow#14 @kmichael08
Got `FileNotFoundError: [Errno 2] No such file or directory: '../data/fit_feat.txt'` in notebook _02_fmri_
