### 2021-12-12 #3895 @kmichael08
Hey @csinva and @jbytecode. I'm happy to tick the remaining boxes and accept the submission.

Thanks for responding to the issues and putting the effort to improve the package.
Both paper and examples are much clearer to me now. Great to hear my suggestions were valuable.
I really like this tool and will use it in my future projects.

Now, I pointed out a few minor details in the issues (without reopening them). I had some minor problems with example notebooks (e.g. imported data is missing in one case) and warnings.
Regardless of that, I accept this with pleasure.

Thanks!
Michał
