### 2021-12-12 Yu-Group/veridical-flow#14 @kmichael08
`%% time` notebook magic doesn't work for me. I needed to remove the space. But maybe it's jupyter version-dependent?
I'm referring to the Ray example.
