### 2021-12-13 5a435c1
fix convert.to_list and add tests
