### 2021-12-14 f8aa42b
Improve lazy and async Vset eval
