### 2021-12-14 dc93067
Handle empty groups in perturbation_stats
