### 2021-12-14 bb71d98
Accept Vset output dict or pandas.DataFrame as first arg to `perturbation_stats`
