### 2021-12-14 81d2d55
remove duplicate compute_interval and change predict_with_uncertainty
