### 2021-12-15 1defb67
Update  API and minor paper edits
