+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'VeridicalFlow: a Python package for building trustworthy data-science pipelines with PCS'
+title: 'VeridicalFlow: a Python package for building trustworthy data science pipelines with PCS'
 tags:
   - python
   - stability
@@ -32,8 +32,8 @@ bibliography: references.bib
 
 # Summary
 
-`VeridicalFlow` is a Python package for simplifying building reproducible and trustworthy data-science pipelines using the PCS framework [@yu2020veridical].
-It provides users with a simple interface for stability analysis, i.e. checking the robustness of results from a data-science pipeline to various judgement calls made during modeling.
+`VeridicalFlow` is a Python package that simplifies building reproducible and trustworthy data science pipelines using the PCS framework [@yu2020veridical].
+It provides users with a simple interface for stability analysis, i.e. checking the robustness of results from a data science pipeline to various judgement calls made during modeling.
 This ensures that arbitrary judgement calls made by data practitioners (e.g. specifying a default imputation strategy) do not dramatically alter the final conclusions made in a modeling pipeline.
 In addition to wrappers facilitating stability analysis, `VeridicalFlow` also automates many cumbersome coding aspects of Python pipelines, including experiment tracking and saving, parallelization, and caching, all through integrations with existing Python packages.
 Overall, the package helps to code using the PCS (predictability-computability-stability) framework, by screening models for predictive performance, helping automate computation, and facilitating stability analysis.
@@ -189,7 +189,7 @@ importances_df = dict_to_df(importances, param_key='out')
 
 # get count, mean, and std of the permutation importances
 perturbation_stats(importances_df, 'preproc', 'RF',
-                   wrt_col='out-importances_mean',
+                   wrt='out-importances_mean',
                    prefix='X', split=True)
 ```
 
