### 2021-12-15 6fcb8e7
fix predict_with_uncertainties
predict_with_uncertainties now returns mean_dict, std_dict computed grouping by
group_by with respect to wrt_col
