### 2021-12-15 7465d03
add predict_with_uncertainties TODOs
