### 2021-12-25 Yu-Group/veridical-flow#27 @richrobe
[JOSS Review] Paper
(as part of: https://github.com/openjournals/joss-reviews/issues/3895)

Just some minor remarks regarding the paper:

* line 2: Explanation of PCS should be introduced when first mentioning it
* line 27: Wrong citation
