### 2021-12-25 Yu-Group/veridical-flow#28 @richrobe
[JOSS Review] Code
(as part of: https://github.com/openjournals/joss-reviews/issues/3895)

When looking at the code there are some points you might want to address which would make usage of the package even better:

* For some functions and classes, docstrings are missing.
* In general, the docstring formatting is inconsistent; some docstrings start with small letters, some with capital letters, some have a Short Summary at the beginning, some don't.
* To easily detect those issues and to further improve code quality I would recommend using a code analysis tool such as [prospector](http://prospector.landscape.io/en/master/), which includes `pylint` for linting and `pep8` for checking against style conventions. This could easily be integrated into your existing workflow and also works well with GitHub Actions. This would also make it easier for contributors to get up to speed quickly with `VeridicalFlow`.
