### 2021-12-25 Yu-Group/veridical-flow#29 @richrobe
[JOSS Review] Example Notebooks
(as part of: https://github.com/openjournals/joss-reviews/issues/3895)

I noticed some issues with the example notebooks that I summarized here:

* In general, it would be very beneficial to have a bit more explanation in the example notebooks in order to better understand the cool features and advantages of `VeridicalFlow`.

* The `00_synthetic_classification.ipynb` notebook throws the following Exception when trying to execute it: `module 'sklearn' has no attribute 'datasets'`. My suggestion would be to add the line `import sklearn.datasets`. Furthermore, the function `sklearn.datasets.load_boston` is deprecated:

    ```
    'load_boston' is deprecated in 1.0 and will be removed in 1.2.
    ```

    Maybe you want to consider using another dataset as an example in order to make it safe for future use?

* Another general improvement suggestion would be to directly include example the notebooks into documentation instead of linking to the notebooks in the GitHub repo.

    Your current solution implies that the example notebooks always have to be fully executed before committing to the repository to include the output in the notebook for. However, this has some drawbacks:

    * In general, it's discouraged to commit notebooks with their output since every execution would change the notebook file
    * And, of course, you would always need to remember to really run the notebooks before committing.

    For building docs with `sphinx` this could be done using [`nbsphinx`](https://nbsphinx.readthedocs.io/en/0.8.7/), To be honest, I don't know how it works with `pdoc`, but I'm sure similar solutions exist?
