### 2021-12-25 #3895 @richrobe
Dear @csinva and @jbytecode. I am very sorry about the late reply, I was on vacation and had GitHub notifications turned off. I was able to finally complete my review now. In summary, I think `VeridicalFlow` seems to be a very nice and useful package! Most checkboxes could be checked off right away but some (small) issues still remain.

Below is the summary of some points that - in my opinion - might help to further improve the package and will help me to check off the remaining boxes for completing the review:

* **Paper**: The paper has a nice structure and reads well, and incorporating the improvement suggestions by @kmichael08 have made it even better! I just have some more minor remarks that are referenced in https://github.com/Yu-Group/veridical-flow/issues/27.

* **Code**: Overall, the code looks well after the suggestions by @kmichael08. In https://github.com/Yu-Group/veridical-flow/issues/28 I collected some possible points of improvements.

* **Examples**: The examples provide a good intuition of how to use `VeridicalFlow`. However, some issues still remain as outlined in https://github.com/Yu-Group/veridical-flow/issues/29.

These are just some thoughts, let me know what you think and we can discuss further about single points! :-) And - again - sorry about the delay.
