### 2022-01-05 3b77b0e
modify evaluate_uncertainty & add docstrings for predict_with_uncertainties
evaluate_uncertainty now takes mean_preds, std_preds (output from `Vset.predict_with_uncertainties` across individual predictions)
- assumes binary classification and computes uncertainty/cumulative accuracy for positive classification (`y=1`)
- output can be plotted as in `00_synthetic_classification.ipynb`
