### 2022-01-05 00923ea
fix imports in 00_synthetic_classification.ipynb & replace boston w/ diabetes dataset #29
