### 2022-01-07 3ad0937
Minor linting fixes
