+++ vflow/convert.py
@@ -16,6 +16,7 @@ from vflow.subkey import Subkey
 from vflow.vfunc import VfuncPromise
 from vflow.vset import PREV_KEY
 
+
 def init_args(args_tuple: Union[tuple, list], names=None):
     """ converts tuple of arguments to a list of dicts
     Params
@@ -29,7 +30,7 @@ def init_args(args_tuple: Union[tuple, list], names=None):
         assert len(names) == len(args_tuple), 'names should be same length as args_tuple'
 
     output_dicts = []
-    for (i, ele) in enumerate(args_tuple):
+    for i in range(len(args_tuple)):
         output_dicts.append({
             (Subkey(names[i], 'init'),): args_tuple[i],
             PREV_KEY: ('init',),
@@ -42,14 +43,14 @@ def s(x):
     """
     if type(x) in [list, tuple]:
         return len(x)
-    else:
-        return x.shape
+    return x.shape
 
 
 def init_step(idx, cols):
     for i in range(idx, len(cols)):
         if cols[i] != 'init':
             return 'init-' + cols[i]
+    return None
 
 
 def dict_to_df(d: dict, param_key=None):
@@ -57,7 +58,7 @@ def dict_to_df(d: dict, param_key=None):
     into a pandas DataFrame, optionally seperating
     parameters in param_key if not None
     """
-    d_copy = {tuple([sk.value for sk in k]): d[k] for k in d if k != PREV_KEY}
+    d_copy = {tuple(sk.value for sk in k): d[k] for k in d if k != PREV_KEY}
     df = pd.Series(d_copy).reset_index()
     if len(d_copy.keys()) > 0:
         key_list = list(d.keys())
@@ -77,10 +78,10 @@ def dict_to_df(d: dict, param_key=None):
                 param_key_cols = [f"{p.split('=')[0]}-{param_key}" for p in param_keys[0]]
                 param_keys = [[s.split('=')[1] for s in t] for t in param_keys]
                 df = df.join(pd.DataFrame(param_keys)).drop(columns=param_key)
-                new_cols = df.columns[:len(cols)-1].tolist() + param_key_cols
+                new_cols = df.columns[:len(cols) - 1].tolist() + param_key_cols
                 df.set_axis(new_cols, axis=1, inplace=True)
                 new_idx = list(range(len(new_cols)))
-                new_idx = new_idx[:param_loc] + new_idx[len(cols)-1:] + new_idx[param_loc:len(cols)-1]
+                new_idx = new_idx[:param_loc] + new_idx[len(cols) - 1:] + new_idx[param_loc:len(cols) - 1]
                 df = df.iloc[:, new_idx]
     return df
 
@@ -95,8 +96,8 @@ def compute_interval(df: DataFrame, d_label, wrt_label, accum=None):
     return df[[wrt_label, d_label]].groupby(wrt_label).agg(accum)
 
 
-def perturbation_stats(data: Union[DataFrame, dict], *group_by: str, wrt: str='out',
-                       func=None, prefix: str=None, split: bool=False):
+def perturbation_stats(data: Union[DataFrame, dict], *group_by: str, wrt: str = 'out',
+                       func=None, prefix: str = None, split: bool = False):
     """Compute statistics for `wrt` in `data`, conditional on `group_by`
 
     Params
@@ -140,37 +141,35 @@ def perturbation_stats(data: Union[DataFrame, dict], *group_by: str, wrt: str='o
         gb = df.groupby(group_by)[wrt]
     else:
         gb = df.groupby(lambda x: True)[wrt]
-    mean_or_std = type(func) is list and 'mean' in func or 'std' in func
[…]
