+++ vflow/vfunc.py
@@ -21,16 +21,14 @@ class Vfunc:
         """
         if hasattr(self.module, 'fit'):
             return self.module.fit(*args, **kwargs)
-        else:
-            return self.module(*args, **kwargs)
+        return self.module(*args, **kwargs)
 
     def transform(self, *args, **kwargs):
         """This function transforms its input in some way
         """
         if hasattr(self.module, 'transform'):
             return self.module.transform(*args, **kwargs)
-        else:
-            return self.module(*args, **kwargs)
+        return self.module(*args, **kwargs)
 
     def __call__(self, *args, **kwargs):
         """This should decide what to call
@@ -61,16 +59,14 @@ class AsyncModule:
         """
         if hasattr(self.module, 'fit'):
             return _remote_fun.remote(self.module.fit, *args, **kwargs)
-        else:
-            return _remote_fun.remote(self.module, *args, **kwargs)
+        return _remote_fun.remote(self.module, *args, **kwargs)
 
     def transform(self, *args, **kwargs):
         """This function transforms its input in some way
         """
         if hasattr(self.module, 'transform'):
             return _remote_fun.remote(self.module.transform, *args, **kwargs)
-        else:
-            return _remote_fun.remote(self.module, *args, **kwargs)
+        return _remote_fun.remote(self.module, *args, **kwargs)
 
     def __call__(self, *args, **kwargs):
         return self.fit(*args, **kwargs)
@@ -110,5 +106,4 @@ class VfuncPromise:
     def __repr__(self):
         if self.called:
             return f'Fulfilled VfuncPromise({self.value})'
-        else:
-            return f'Unfulfilled VfuncPromise(func={self.vfunc}, args={self.args})'
+        return f'Unfulfilled VfuncPromise(func={self.vfunc}, args={self.args})'
