+++ vflow/pipeline.py
@@ -3,7 +3,6 @@
 import itertools
 
 import joblib
-import matplotlib.pyplot as plt
 import networkx as nx
 import pandas as pd
 
@@ -41,7 +40,7 @@ class PCSPipeline:
             except AttributeError:
                 step_name = f'Step {i}'
             print(step_name)
-            outputs, fitted_step = run_step_cached(step, *args, **kwargs)
+            _, fitted_step = run_step_cached(step, *args, **kwargs)
             self.steps[i] = fitted_step
 
     def __getitem__(self, i):
@@ -60,11 +59,10 @@ class PCSPipeline:
                                    for i, mod in enumerate(step)])
             indexes = list(itertools.product(*name_lists))
             return pd.DataFrame(indexes, columns=[step.name for step in self.steps])
-        else:
-            for step in self.steps:
-                name_lists.append([f'{step.name}_{i}_{str(mod)[:8]}'
-                                   for i, mod in enumerate(step)])
-            return list(itertools.product(*name_lists))
+        for step in self.steps:
+            name_lists.append([f'{step.name}_{i}_{str(mod)[:8]}'
+                               for i, mod in enumerate(step)])
+        return list(itertools.product(*name_lists))
 
 
 def build_graph(node, draw=True):
@@ -107,11 +105,11 @@ def build_graph(node, draw=True):
         G: nx.Digraph()
         """
         # base case: reached starting node
-        if type(node) is str:
+        if isinstance(node, str):
             return G
 
         # initial case: starting at dict
-        elif type(node) is dict:
+        if isinstance(node, dict):
             s_node = 'End'
             nodes_prev = node[PREV_KEY]
             G.add_edge(nodes_prev[0], s_node)
@@ -121,7 +119,7 @@ def build_graph(node, draw=True):
             return G
 
         # main case: at a moduleset
-        elif 'Vset' in str(type(node)):
+        if 'Vset' in str(type(node)):
             if hasattr(node, PREV_KEY):
                 nodes_prev = getattr(node, PREV_KEY)
                 for node_prev in nodes_prev:
@@ -130,13 +128,15 @@ def build_graph(node, draw=True):
             return G
 
         # nested prev key case
-        elif type(node) is tuple:
+        if isinstance(node, tuple):
             func_node = unnest_node(node[0])
             G = build_graph_recur(func_node, G)
             for arg_node in node[1:]:
                 G.add_edge(unnest_node(arg_node), func_node)
                 G = build_graph_recur(arg_node, G)
             return G
+
+        return G
 
     G = nx.DiGraph()
     G = build_graph_recur(node, G)
