### 2022-01-07 c1f4acb
Add pylint to python-package github action
