+++ setup.py
@@ -30,6 +30,7 @@ setuptools.setup(
     extras_require={
         'dev': [
             'pytest',
+            'pylint==2.12.2',
             'tqdm',
             'scikit-learn >=0.23.0',  # 0.23+ only works on py3.6+)
         ],
