+++ .github/workflows/python-package.yml
@@ -23,6 +23,9 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install .[dev] --no-cache-dir
+      - name: Lint with pylint
+        run: |
+          pylint vflow *.py --rcfile=.pylintrc
       - name: Test with pytest
         run: |
           pytest
