### 2022-01-07 e04960a
Fix bug in VfuncPromise when combining lazy and async evaluation
When Vset's transform, predict, or predict_proba are called lazily,
the resulting output dict values are VfuncPromises, but if the
underlying Vfunc is async (via the AsyncModule class) then the result
of calling the VfuncPromise is a ray.ObjectRef. Thus, we first need to
get the result from Ray's object store before calling the transform,
predict, or predict_proba method.
