### 2022-01-07 94aae4f
Additional minor linting fixes
