### 2022-01-08 537a509
evaluate_uncertainty supports group_by
fix evaluate_uncertainty to support returning list of
uncertainty/cumulative accuracy per each item in mean_preds/std_preds
- mean_preds/std_preds keys must be identical
- binary predictions computed for `y=1`
