### 2022-01-09 Yu-Group/veridical-flow#27 @csinva
Thanks for catching these, we just updated the paper to address these remarks!
