### 2022-01-10 8db4d85
improved and standardized docstrings
addresses #28
