+++ vflow/vset.py
@@ -20,29 +20,29 @@ class Vset:
                  lazy: bool = False, cache_dir: str = None,
                  tracking_dir: str = None):
         """
-        todo: include prev and next and change functions to include that.
-        Params
-        -------
+        Parameters
+        ----------
         name: str
-            name of this moduleset
+            Name of this Vset.
         modules: list or dict
-            dictionary of functions that we want to associate with
+            Dictionary of functions that we want to associate with
         module_keys: list (optional)
-            list of names corresponding to each module
+            List of names corresponding to each module
         is_async: bool (optional)
-            if True, modules are computed asynchronously
+            If True, `modules` are computed asynchronously
         output_matching: bool (optional)
-            if True, then output keys from this Vset will be matched when used
+            If True, then output keys from this Vset will be matched when used
             in other Vsets
         lazy: bool (optional)
-            if True, then modules are evaluated lazily, i.e. outputs contain a
-            promise
+            If True, then modules are evaluated lazily, i.e. outputs are `VfuncPromise`
         cache_dir: str (optional)
-            if provided, do caching and use cache_dir as the data store for
-            joblib.Memory
+            If provided, do caching and use `cache_dir` as the data store for
+            `joblib.Memory`.
         tracking_dir: str (optional)
-            if provided, use the mlflow.tracking api to log outputs as metrics
-            with params determined by input keys
+            If provided, use the `mlflow.tracking` api to log outputs as metrics
+            with params determined by input keys.
+        
+        .. todo:: include prev and next and change functions to include that.
         """
         self.name = name
         self._fitted = False
@@ -88,6 +88,30 @@ class Vset:
                 self.modules[k] = Vfunc(k[0], v)
 
     def _apply_func(self, out_dict: dict = None, *args):
+        """Apply functions in out_dict to combined args dict
+
+        Optionally logs output Subkeys and values as params and metrics using
+        `MlflowClient` if this Vset has a `_tracking_dir`.
+
+        Parameters
+        ----------
+        *args: dict
+            Takes multiple dicts and combines them into one.
+            Then runs modules on each item in combined dict.
+        out_dict: dict (optional), default None
+            The dictionary to pass to the matching function. If None, defaults to self.modules.
+
+        Returns
+        -------
+        out_dict: dict
+            Dictionary with items being determined by functions in module set.
+            Functions and input dictionaries are currently matched using a cartesian matching format.
+        
+        Examples
+        --------
+        >>> modules, data = {LR : logistic}, {train_1 : [X1,y1], train2 : [X2,y2]}
+        {(train_1, LR) : fitted logistic, (train_2, LR) :  fitted logistic}
+        """
         if out_dict is None:
             out_dict = deepcopy(self.modules)
 
@@ -133,7 +157,7 @@ class Vset:
         return out_dict
 
     def fit(self, *args):
-        """
+        """Fits to args using `_apply_func`
         """
         out_dict = {}
         for k, v in self.modules.items():
@@ -152,7 +176,7 @@ class Vset:
         return self.fit(*args).transform(args[0])
 
     def transform(self, *args):
-        """
+        """Transforms args using `_apply_func`
         """
         if not self._fitted:
             raise AttributeError('Please fit the Vset object before calling the transform method.')
@@ -163,6 +187,8 @@ class Vset:
         return self._apply_func(out_dict, *args)
 
     def predict(self, *args):
+        """Predicts args using `_apply_func`
+        """
         if not self._fitted:
             raise AttributeError('Please fit the Vset object before calling the predict method.')
[…]
