+++ vflow/subkey.py
@@ -2,13 +2,14 @@ class Subkey:
 
     def __init__(self, value, origin: str, output_matching: bool = False):
         """
-        Params
-        -------
+        Parameters
+        ----------
         value:
             subkey value corresponding to a Vset module
         origin: str
             name of the origin Vset of this Subkey
-        output_matching: inherited from the Vset where the Subkey is created
+        output_matching: bool (optional), default False
+            inherited from the Vset where the Subkey is created
         """
         self.value = value
         self.origin = origin
@@ -18,9 +19,13 @@ class Subkey:
         self._sep_dicts_id = None
 
     def is_matching(self):
+        """Checks if subkey should be matched in other Vsets
+        """
         return self._output_matching or self._sep_dicts_id is not None
 
     def matches_sep_dict_id(self, o: object):
+        """Helper to match Subkey by _sep_dict_id
+        """
         if isinstance(o, self.__class__):
             return self._sep_dicts_id is not None \
                    and self._sep_dicts_id == o._sep_dicts_id
@@ -29,8 +34,7 @@ class Subkey:
     def matches(self, o: object):
         """When Subkey matching is required, determines if this Subkey is compatible
         with another, meaning that the origins and values match, and either the
-        _sep_dicts_ids match or both Subkeys have _output_matching True.
-
+        _sep_dicts_id matches or both Subkeys have _output_matching True.
         """
         if isinstance(o, self.__class__):
             # they're both matching
@@ -62,19 +66,17 @@ class Subkey:
         return True
 
     def __eq__(self, o: object):
-        """
-        Mainly used for testing purposes.
+        """Mainly used for testing purposes.
         """
         if isinstance(o, self.__class__):
             # value and origins match
             return self.value == o.value and self.origin == o.origin
         return False
 
-    def __repr__(self) -> str:
+    def __repr__(self):
         return str(self.value)
 
     def __hash__(self):
-        """
-        Mainly used for testing purposes.
+        """Mainly used for testing purposes.
         """
         return hash(self.value) ^ hash(self.origin) ^ hash(self._output_matching)
