+++ vflow/pipeline.py
@@ -12,18 +12,14 @@ from vflow.vset import PREV_KEY
 
 class PCSPipeline:
     def __init__(self, steps=None, cache_dir=None):
-        """Helper function that just calls build_graph_recur with an empty graph
-        Params
-        ------
+        """
+        Parameters
+        ----------
         steps: list
             a list of Vset instances
         cache_dir: str, default=None
             The directory to use as data store by joblib. If None, won't do
             caching.
-
-        Returns
-        -------
-        G: nx.Digraph()
         """
         if steps is None:
             steps = []
@@ -69,8 +65,8 @@ class PCSPipeline:
 
 def build_graph(node, draw=True):
     """Helper function that just calls build_graph_recur with an empty graph
-    Params
-    ------
+    Parameters
+    ----------
     node: dict or Vset
 
     Returns
@@ -97,8 +93,8 @@ def build_graph(node, draw=True):
 
     def build_graph_recur(node, G):
         """Builds a graph up using __prev__ and PREV_KEY pointers
-        Params
-        ------
+        Parameters
+        ----------
         node: str, dict, Vset, or tuple
         G: nx.Digraph()
 

[…]
