### 2022-01-10 365850a
update docstrings and rebuild docs
