### 2022-01-10 301538d
Add with_uncertainty param to Vset.predit and Vset.predict_proba
