### 2022-01-10 7edc0b9
Return accuracy from cum_acc_by_uncertainty
Renames evaluate_uncertainty to cum_acc_by_uncertainty and adds some
additional tests.
