### 2022-01-11 10d9dcd
refactor convert.py to utils.py
- moved init_args, dict_to_df, perturbation_stats & cum_acc_by_uncertainty to helpers
- renamed convert to utils
