+++ tests/test_utils.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 from numpy.testing import assert_equal
 
-from vflow.convert import *
+from vflow.utils import *
 from vflow.subkey import Subkey as sm
 import pandas as pd
 
@@ -1009,7 +1009,7 @@ class TestApplyModules:
         result_dict = apply_modules(*in_dicts)
         assert_equal(result_dict, out_dict)
 
-class TestConvert:
+class TestUtils:
 
     def test_to_list(self):
         assert to_list((['x1', 'x2', 'x3'], ['y1', 'y2', 'y3'])) == [['x1', 'y1'], ['x2', 'y2'], ['x3', 'y3']]
@@ -1020,114 +1020,3 @@ class TestConvert:
         assert to_list(('x1', 'x2', 'x3', 'y1', 'y2', 'y3')) == [['x1', 'y1'], ['x2', 'y2'], ['x3', 'y3']]
         with pytest.raises(ValueError):
             to_list(('x1', 'x2', 'x3', 'y1', 'y2'))
-
-    def test_dict_to_df(self):
-        in_dict_1 = {(sm('X_train', 'init'), sm('feat_extract_0', 'feat_extract'), 
-            sm('y_train', 'init'), sm('DT', 'modeling'), sm('acc', 'metrics')): 0.9,
-            (sm('X_train', 'init'), sm('feat_extract_1', 'feat_extract'), 
-            sm('y_train', 'init'), sm('DT', 'modeling'), sm('acc', 'metrics')): 0.95}
-        out_df_1 = pd.DataFrame(data={'init-feat_extract': ['X_train', 'X_train'],
-                                'feat_extract': ['feat_extract_0', 'feat_extract_1'],
-                                'init-modeling': ['y_train', 'y_train'],
-                                'modeling': ['DT', 'DT'],
-                                'metrics': ['acc', 'acc'],
-                                'out': [0.9, 0.95]})
-        in_dict_2 = {(sm('X_train', 'init'), sm('sample_0', 'sample'), sm('y_train', 'init'), 
-                    sm(('k=10', 'e=1e-3'), 'modeling'), sm('s_0', 'stability')): 0.333,
-                    (sm('X_train', 'init'), sm('sample_0', 'sample'), sm('y_train', 'init'), 
-                    sm(('k=10', 'e=1e-5'), 'modeling'), sm('s_0', 'stability')): 0.452}
-        out_df_2 = pd.DataFrame(data={'init-sample': ['X_train', 'X_train'],
-                                'sample': ['sample_0', 'sample_0'],
-                                'init-modeling': ['y_train', 'y_train'],
-                                'k-modeling': ['10', '10'],
-                                'e-modeling': ['1e-3', '1e-5'],
-                                'stability': ['s_0', 's_0'],
-                                'out': [0.333, 0.452]})
-        in_dict_3 = {(sm('X_train', 'init'), sm('sample_0', 'sample'), sm('y_train', 'init'),
-                    sm(('k=10', 'e=1e-3'), 'modeling'), sm('s_0', 'stability')): [0.333, 0.222],
-                    (sm('X_train', 'init'), sm('sample_0', 'sample'), sm('y_train', 'init'),
-                     sm(('k=10', 'e=1e-5'), 'modeling'), sm('s_0', 'stability')): [0.452, 0.322]}
-        out_df_3 = pd.DataFrame(data={'init-sample': ['X_train', 'X_train'],
-                                'sample': ['sample_0', 'sample_0'],
-                                'init-modeling': ['y_train', 'y_train'],
[…]
