+++ vflow/utils.py
@@ -7,7 +7,6 @@ from uuid import uuid4
 import numpy as np
 import pandas as pd
 import numpy as np
-from pandas import DataFrame
 
 import ray
 from ray.remote_function import RemoteFunction as RayRemoteFun
@@ -17,28 +16,6 @@ from vflow.subkey import Subkey
 from vflow.vfunc import VfuncPromise
 from vflow.vset import PREV_KEY
 
-def init_args(args_tuple: Union[tuple, list], names=None):
-    """Converts tuple of arguments to a list of dicts
-    
-    Parameters
-    ----------
-    names: list-like (optional), default None
-        given names for each of the arguments in the tuple
-    """
-    if names is None:
-        names = ['start'] * len(args_tuple)
-    else:
-        assert len(names) == len(args_tuple), 'names should be same length as args_tuple'
-
-    output_dicts = []
-    for (i, ele) in enumerate(args_tuple):
-        output_dicts.append({
-            (Subkey(names[i], 'init'),): args_tuple[i],
-            PREV_KEY: ('init',),
-        })
-    return output_dicts
-
-
 def s(x):
     """Gets shape of a list/tuple/ndarray
     """
@@ -46,7 +23,6 @@ def s(x):
         return len(x)
     else:
         return x.shape
-
 
 def init_step(idx, cols):
     """Helper function to find init suffix
@@ -63,83 +39,6 @@ def init_step(idx, cols):
         if cols[i] != 'init':
             return 'init-' + cols[i]
 
-
-def dict_to_df(d: dict, param_key=None):
-    """Converts a dictionary with tuple keys
-    into a pandas DataFrame, optionally seperating
-    parameters in `param_key` if not None
-
-    Parameters
-    ----------
-    d: dict
-        Output dictionary with tuple keys from a Vset.
-    param_key: str (optional), default None
-        Name of parameter to seperate into multiple columns.
-
-    Returns
-    -------
-    df: pandas.DataFrame
-        A DataFrame with `d` tuple keys seperated into columns.
-    """
-    d_copy = {tuple([sk.value for sk in k]): d[k] for k in d if k != PREV_KEY}
-    df = pd.Series(d_copy).reset_index()
-    if len(d_copy.keys()) > 0:
-        key_list = list(d.keys())
-        subkey_list = key_list[0] if key_list[0] != PREV_KEY else key_list[1]
-        cols = [sk.origin for sk in subkey_list] + ['out']
-        # set each init col to init-{next_module_set}
-        cols = [c if c != 'init' else init_step(idx, cols) for idx, c in enumerate(cols)]
-        df.set_axis(cols, axis=1, inplace=True)
-        if param_key:
-            param_keys = df[param_key].tolist()
-            if param_key == 'out' and hasattr(param_keys[0], '__iter__'):
-                param_df = pd.DataFrame(param_keys)
-                param_df.columns = [f'{param_key}-{col}' for col in param_df.columns]
-                df = df.join(param_df)
-            else:
-                param_loc = df.columns.get_loc(param_key)
-                param_key_cols = [f"{p.split('=')[0]}-{param_key}" for p in param_keys[0]]
-                param_keys = [[s.split('=')[1] for s in t] for t in param_keys]
-                df = df.join(pd.DataFrame(param_keys)).drop(columns=param_key)
-                new_cols = df.columns[:len(cols)-1].tolist() + param_key_cols
-                df.set_axis(new_cols, axis=1, inplace=True)
-                new_idx = list(range(len(new_cols)))
-                new_idx = new_idx[:param_loc] + new_idx[len(cols)-1:] + new_idx[param_loc:len(cols)-1]
-                df = df.iloc[:, new_idx]
-    return df
-
[…]
