+++ vflow/__init__.py
@@ -4,5 +4,5 @@
 from .vfunc import *
 from .vset import *
 from .pipeline import *
-from .convert import init_args, dict_to_df, compute_interval, perturbation_stats
+from .utils import *
 from .helpers import *
