+++ vflow/vset.py
@@ -9,7 +9,8 @@ import numpy as np
 import ray
 from mlflow.tracking import MlflowClient
 
-from vflow.helpers import perturbation_stats
+from vflow.utils import *
+from vflow.helpers import perturbation_stats, dict_to_df
 from vflow.subkey import Subkey
 from vflow.vfunc import Vfunc, AsyncModule, VfuncPromise, _remote_fun
 
