### 2022-01-11 2403250
More minor linting fixes
