+++ vflow/helpers.py
@@ -7,7 +7,7 @@ from typing import Union
 import pandas as pd
 import numpy as np
 
-from vflow.utils import init_step, dict_keys, dict_data
+from vflow.utils import dict_keys, dict_data
 from vflow.vfunc import Vfunc
 from vflow.vset import Vset, Subkey, PREV_KEY, FILTER_PREV_KEY
 
@@ -31,50 +31,6 @@ def init_args(args_tuple: Union[tuple, list], names=None):
             PREV_KEY: ('init',),
         })
     return output_dicts
-
-def dict_to_df(d: dict, param_key=None):
-    """Converts a dictionary with tuple keys
-    into a pandas DataFrame, optionally seperating
-    parameters in `param_key` if not None
-
-    Parameters
-    ----------
-    d: dict
-        Output dictionary with tuple keys from a Vset.
-    param_key: str (optional), default None
-        Name of parameter to seperate into multiple columns.
-
-    Returns
-    -------
-    df: pandas.DataFrame
-        A DataFrame with `d` tuple keys seperated into columns.
-    """
-    d_copy = {tuple([sk.value for sk in k]): d[k] for k in d if k != PREV_KEY}
-    df = pd.Series(d_copy).reset_index()
-    if len(d_copy.keys()) > 0:
-        key_list = list(d.keys())
-        subkey_list = key_list[0] if key_list[0] != PREV_KEY else key_list[1]
-        cols = [sk.origin for sk in subkey_list] + ['out']
-        # set each init col to init-{next_module_set}
-        cols = [c if c != 'init' else init_step(idx, cols) for idx, c in enumerate(cols)]
-        df.set_axis(cols, axis=1, inplace=True)
-        if param_key:
-            param_keys = df[param_key].tolist()
-            if param_key == 'out' and hasattr(param_keys[0], '__iter__'):
-                param_df = pd.DataFrame(param_keys)
-                param_df.columns = [f'{param_key}-{col}' for col in param_df.columns]
-                df = df.join(param_df)
-            else:
-                param_loc = df.columns.get_loc(param_key)
-                param_key_cols = [f"{p.split('=')[0]}-{param_key}" for p in param_keys[0]]
-                param_keys = [[s.split('=')[1] for s in t] for t in param_keys]
-                df = df.join(pd.DataFrame(param_keys)).drop(columns=param_key)
-                new_cols = df.columns[:len(cols)-1].tolist() + param_key_cols
-                df.set_axis(new_cols, axis=1, inplace=True)
-                new_idx = list(range(len(new_cols)))
-                new_idx = new_idx[:param_loc] + new_idx[len(cols)-1:] + new_idx[param_loc:len(cols)-1]
-                df = df.iloc[:, new_idx]
-    return df
 
 def build_vset(name: str, obj, param_dict=None, *args, reps: int = 1,
                is_async: bool = False, output_matching: bool = False,
@@ -220,91 +176,6 @@ def filter_vset_by_metric(metric_dict: dict, vset: Vset, *vsets: Vset, n_keep: i
         return vsets[0]
     else:
         return vsets
-
-def perturbation_stats(data: Union[pd.DataFrame, dict], *group_by: str, wrt: str='out',
-                       func=None, prefix: str=None, split: bool=False):
-    """Compute statistics for `wrt` in `data`, conditional on `group_by`
-
-    Parameters
-    ----------
-    data: Union[pandas.DataFrame, dict]
-        DataFrame, as from calling `dict_to_df` on an output dict from a Vset,
-        or the output dict itself.
-    *group_by: str
-        Vset names in `data` to group on. If none provided, treats everything as one big
-        group.
-    wrt: str (optional)
[…]
