### 2022-01-11 ce49d9d
Fix issue with how filter_vset_by_metric gets mlflow tracking_dir
