### 2022-01-11 35ec918
Move pylint to end of github actions workflow
