### 2022-01-11 7364b1a
More minor linting fixes
