### 2022-01-11 57934fc
Make test_lazy_eval backwards compatible
