### 2022-01-11 Yu-Group/veridical-flow#28 @csinva
Thanks for this comment, we've gone through and improved the docstrings / linting throughout. We've also added pylint via a Github Action, so that it should be checked on every push!
