### 2022-01-11 Yu-Group/veridical-flow#29 @csinva
Thanks for these comments! We've added more detail and some fixes to the notebook examples!

We've also rendered the notebooks + their outputs directly to html (using `nbconvert`), and set up a script to this on every new push. (For example, see the links under the [Documentation tab](https://vflow.csinva.io/#:~:text=subsampling%20or%20model.-,Documentation,-See%20the%20docs)).
