### 2022-01-11 #3895 @csinva
Thanks for all your thoughtful feedback @richrobe. We’ve just finished going through and applying all of your recommendations resulting in a much-improved package / paper!

We’ve responded to each of the issues individually in their own threads and improved the package accordingly. Please let us know if there's something you feel is still an issue!

Best,
Chandan
