### 2022-01-12 #3895 @richrobe
Hi @csinva,
awesome! Thanks for your effort put into responding to the issues and to improve the package. With that, I'm happy to tick off the remaining checkboxes and accept the submission!

Best,
Robert
