### 2022-01-12 6c61d23
format bib entry
