+++ references.bib
@@ -34,9 +34,9 @@
 }
 
 @inproceedings{moritz2018ray,
-  title={Ray: A distributed framework for emerging $\{$AI$\}$ applications},
+  title={Ray: A distributed framework for emerging AI applications},
   author={Moritz, Philipp and Nishihara, Robert and Wang, Stephanie and Tumanov, Alexey and Liaw, Richard and Liang, Eric and Elibol, Melih and Yang, Zongheng and Paul, William and Jordan, Michael I and others},
-  booktitle={13th $\{$USENIX$\}$ Symposium on Operating Systems Design and Implementation ($\{$OSDI$\}$ 18)},
+  booktitle={13th USENIX Symposium on Operating Systems Design and Implementation (OSDI 18)},
   pages={561--577},
   year={2018}
 }
