### 2022-01-12 #3895 @csinva
Thanks for making these changes @jbytecode! I've accepted them + added a reference to table 1 in the manuscript.

Best,
Chandan
