### 2022-01-12 #3895 @jbytecode
It seems the table is saved as a figure, however in fact, it is a table. Please design it as a table and give a header and caption then mention this table in text (e.g. In Table 1...)
