### 2022-01-12 #3895 @jbytecode
if you are not familiar with the table stuff in markdown, here is an example manuscript that uses tables: https://github.com/jbytecode/LinRegOutliers/tree/master/paper
