### 2022-01-12 #3895 @csinva
Thanks for the link to the example manuscript! Just reformatted it as a markdown table.
