### 2022-01-12 #3895 @csinva
Ah yes,

`10.5281/zenodo.5834875` is for all versions "This DOI represents all versions, and will always resolve to the latest one"

`10.5281/zenodo.5834876` is for this version specifically, not sure which one is preferred
