### 2022-01-12 ca95826
minor changes in paper
