### 2022-01-12 #3895 @danielskatz
@csinva - I've suggested some changes in https://github.com/Yu-Group/veridical-flow/pull/38.  Please merge this, or let me know what you disagree with.

In addition:
* your first reference is unclear - can you add a DOI, URL, etc so that the reader will be able to follow this reference to something?
* for the pandas reference, the authors don't match those in the DOI - please fix this.

Once these changes are made, please let me know and we will proceed.
