+++ references.bib
@@ -23,7 +23,36 @@
 
 @article{mckinney2011pandas,
   title={{p}andas: a foundational Python library for data analysis and statistics},
-  author={McKinney, Wes and others},
+    author       = {Jeff Reback and
+                  Wes McKinney and
+                  jbrockmendel and
+                  Joris Van den Bossche and
+                  Tom Augspurger and
+                  Phillip Cloud and
+                  gfyoung and
+                  Sinhrks and
+                  Adam Klein and
+                  Matthew Roeschke and
+                  Jeff Tratner and
+                  Chang She and
+                  William Ayd and
+                  Simon Hawkins and
+                  Terji Petersen and
+                  Jeremy Schendel and
+                  Andy Hayden and
+                  Marc Garcia and
+                  Vytautas Jancauskas and
+                  MomIsBestFriend and
+                  Pietro Battiston and
+                  Skipper Seabold and
+                  chris-b1 and
+                  h-vetinari and
+                  Stephan Hoyer and
+                  Wouter Overmeire and
+                  alimcmaster1 and
+                  Mortada Mehyar and
+                  Christopher Whelan and
+                  Thomas Kluyver},
   journal={Python for high performance and scientific computing},
   volume={14},
   number={9},
@@ -136,5 +165,6 @@
 
 @misc{hagbergnetworkx,
   title={NetworkX: Network Analysis with {P}ython},
-  author={Hagberg, Aric and Conway, Drew}
+  author={Hagberg, Aric and Conway, Drew},
+  url={https://networkx.org/},
 }
