### 2022-01-12 #3895 @csinva
Thanks @danielskatz, I've merged in the changes + updated the two references you mentioned!

Also a huge thank you to @jbytecode for doing a wonderful job serving as editor of this submission and heartfelt thanks to @richrobe and @kmichael08 for their thoughtful reviews.
