### 2022-01-13 #3895 @jpdunc23
@danielskatz I noticed an issue with the affiliations that we just fixed via https://github.com/Yu-Group/veridical-flow/commit/fe106489b3c888991c788c3f01b5461e096f29fe. Is it too late to update the version at https://doi.org/10.21105/joss.03895? Sorry for the trouble.
