### 2022-01-23 #3895 @jpdunc23
> @jpdunc23 – I've updated the paper based on your changes. The updated PDF might take 24 hours to show properly due to caching on the JOSS site.

Thank you!
